Fill properly source_outcomes and event_outcomes.
Also fix messages locking mechanism which now is enabled for globally
sequential handling. This is very slow though and we should find a way
to run handler concurrently. For now stick with this in order to allow
source_outcomes to be properly filled (market_id, event_id) and let the
code populate event_outcomes as needed.